Add step to input types number and range (#93)
Input component was missing the step functionality for input of type
"number" and "range", which we now require for some pages in
AdvisorySuite. This implementation is trivial because step is already a
browser supported functionality, we are just making sure only input of
the mentioned types can pass it as prop.